+++ ReleaseNotes/release_notes.md
@@ -17,6 +17,7 @@
   1. add `StdMoment` to `ostap.stats.moments`
   1. sdd `std_moment` method for `PDF`
   1. mase a bypass for long standing issue with segfaults from `RooFitResult::globalCorr`
+  1. Add Novosiboirsk function and PDF 
   
 ## Backward incompatible: